pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^1.7.3"
aiocron = "^1.8"
ics = "^0.7"
httpx = "^0.21.3"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
pdbpp = "^0.10.3"

[tool.isort]
profile = "black"
